+++ changelog.md
@@ -4,9 +4,16 @@
 ## Master
 
 * Add missing documentation warning.
+* Add new experimental producer API
+* Add initial headers support
+
+<a name="0.16.0"></a>
+## 0.16.0 (2018-05-20)
+
+* Upgrade librdkafka to 0.11.4
 
 <a name="0.15.0"></a>
-## 0.15.0 (2018-03015)
+## 0.15.0 (2018-03-15)
 
 * Added iterator interface to the `BaseConsumer`.
 * Change timeout to more rust-idiomatic `Option<Duration>`.